+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - [#33](https://github.com/bobozaur/sqlx-exasol/pull/33): Avoid nested boxing in Executor impl
+- [#31](https://github.com/bobozaur/sqlx-exasol/pull/31):
+  - Renamed `ExaConnection::socket_addr` to `ExaConnection::server`
 
 ## [0.8.6] - 2025-06-03